# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SecurityGuard Pro API
  description: |
    Comprehensive API documentation for SecurityGuard Pro - A security guard management system with real-time monitoring, geofencing, attendance tracking, and billing capabilities.
    
    ## Features
    - **Guard Management**: Create, update, and manage security guards
    - **Real-time Location Tracking**: GPS-based location monitoring with geofencing
    - **Attendance System**: Automated attendance tracking with facial recognition
    - **Geofencing**: Create and monitor virtual boundaries
    - **Alert Management**: Real-time alerts and notifications
    - **Billing & Subscriptions**: Integrated payment processing with Stripe
    - **Multi-tenancy**: Support for multiple organizations
    - **Mobile Support**: Offline data synchronization for mobile apps
    
    ## Authentication
    All API endpoints require authentication using JWT tokens obtained through the authentication endpoints.
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per 15-minute window per IP address.
    
    ## Base URL
    - Production: `https://api.securityguardpro.com/v1`
    - Development: `http://localhost:3000/api`
  version: 1.0.0
  contact:
    name: SecurityGuard Pro API Support
    email: api-support@securityguardpro.com
    url: https://securityguardpro.com/docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Geofence:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        radius:
          type: number
          format: double
        type:
          type: string
          enum: [CIRCLE, POLYGON]
        coordinates:
          type: array
          items:
            type: array
            items:
              type: number
              format: double
        isActive:
          type: boolean
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        domain:
          type: string
        settings:
          type: object
        subscription:
          $ref: '#/components/schemas/Subscription'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        plan:
          type: string
          enum: [BASIC, PROFESSIONAL, ENTERPRISE]
        status:
          type: string
          enum: [ACTIVE, CANCELLED, EXPIRED, PENDING]
        stripeSubscriptionId:
          type: string
        currentPeriodStart:
          type: string
          format: date-time
        currentPeriodEnd:
          type: string
          format: date-time
        trialEnd:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /geofences:
    get:
      tags:
        - Geofences
      summary: Get Geofences
      description: Retrieve a list of geofences
      responses:
        '200':
          description: Geofences retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Geofence'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Geofences
      summary: Create Geofence
      description: Create a new geofence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - latitude
                - longitude
                - radius
                - type
              properties:
                name:
                  type: string
                description:
                  type: string
                latitude:
                  type: number
                  format: double
                longitude:
                  type: number
                  format: double
                radius:
                  type: number
                  format: double
                type:
                  type: string
                  enum: [CIRCLE, POLYGON]
                coordinates:
                  type: array
                  items:
                    type: array
                    items:
                      type: number
                      format: double
                isActive:
                  type: boolean
                  default: true
      responses:
        '201':
          description: Geofence created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  # Attendance Endpoints
...
  /tenants:
    get:
      tags:
        - Tenants
      summary: Get Tenants
      description: Retrieve a list of tenants (admin only)
      responses:
        '200':
          description: Tenants retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tenant'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Tenants
      summary: Create Tenant
      description: Create a new tenant (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - domain
              properties:
                name:
                  type: string
                domain:
                  type: string
                settings:
                  type: object
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...